// src/Client/Logistics.Angular/projects/shared/src/lib/api/generated/models/document-type.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export type DocumentType = 'bill_of_lading' | 'proof_of_delivery' | 'invoice' | 'receipt' | 'contract' | 'insurance_certificate' | 'photo' | 'driver_license' | 'vehicle_registration' | 'identity_document' | 'pickup_inspection' | 'delivery_inspection' | 'other' | 'dot_inspection' | 'title_certificate' | 'lease_agreement' | 'maintenance_record' | 'annual_inspection';